Forbid ELSEIF...ELSE...IF sequences.
lang.g can never generate these, but compiler.g would have generated crazy
op sequences for them:

    ^(ENDIF
        ^(ELSEIF elseif_cond elseif_body)
        ^(ELSE else_body)
        ^(IF if_cond if_body)
    )

becomes

    if_cond
    ^(OP_IF $ELSE)
    if_body
    ^(OP_JUMP $ENDIF)
    $ELSE:
    else_body
    elseif_cond
    ^(OP_EIF $ENDIF)
    elseif_body
    ^(OP_JUMP $ENDIF)
    $ENDIF: ...rest of program...

which isn't right in the head.

(If OP_EIF was just OP_IF, this would look a lot like, but not identical to,
two if-statements with some intervening code. Note the target of the jump at
the end of if_body.)

We now only permit ^(ELSE else_body) as the *first* child of an ^(ENDIF ...)
tree.