Added support for WHILE statements.
This also adds support for bare identifiers as expressions, to exercise cases
like `while identifier_while (identifier_condition) identifier_body;
endwhile`. To make this parse cleanly, the loop identifier is tagged in the
parse tree as a ^(LOOP_TAG ...) subtree. This avoids ambiguous parse trees
such as

    ^(ENDWHILE
        identifier_while
        identifier_condition
        identifier_body
    )

I've called it LOOP_TAG in the hopes that the same construct can apply to
other kinds of loops.