Compile FORK statements.
Fork statements result in a secondary code vector (one per fork statement), so
I've introduced a new top-level element in the compiler.g output: FORK_VECTOR.
The intent is that it contains the ops destined for a new vector. It also
appears in OP_FORK and OP_FORK_WITH_ID ops, as a way of referencing the new
fork vector, as with LABEL in jump instructions.