Compile try/except statements.
Unlike if/elseif/else statements, the parse tree is the right way around. (We
don't have to generate a linked list.) However, the first except in a try is
special-cased to generate different op sequences so that the *last* except
doesn't end with a no-op jump. This differs from bare-if statements (which do
end with a no-op jump) intentionally.