Wrote a little virtualenv-alike for MySQL.
This is meant to ease spinning up temporary database _servers_ for development
and testing. I use it to test database migrations, in isolation, by ensuring
that the test starts from a known clean state: a brand-new server.

The newly-created server comes up on a randomly-selected port between `10000`
and `42767`, inclusive, with all of its data and configuration contained in
the environment directory. The initial configuration file contains what I
think of as "good ideas" for MySQL servers; there's a command line flag for
specifying an alternate config file if you disagree with me. Most of the
defaults are non-intrusive things like collation or storage engine choices.

(I stopped short of setting a restrictive SQL mode, but I really
thought about it.)

Much like virtualenv, you "activate" a MySQL environment by sourcing its
`activate` script. This provides a corresponding `deactivate-mysql` command
while the environment is active, _and starts (and stops) MySQL_ for you. The
basic command line tools `mysql`, `mysqladmin`, and `mysqldump` are
automatically configured to talk to the environment's server, as root by
default.