Created a simple bot framework.
Alicorn is designed to run "handlers", which are the blocks of code that make
it do interesting things like respond to messages or log your chat for you.
These handlers are enumerated in a YAML file (an example is included) that
also determines what IRC network the bot connects to.

I've included some simple support (see `alicorn/io.py`) for sharing files
(logs, for example) between multiple bots.