Added support for Nickserv identification.
On connection startup, this module sends a privmsg to nickserv (specifically,
to the nick given in the `nickserv` config option, defaulting to "NickServ"
per Freenode) identifying the bot. The message has one of the follow forms:

* `IDENTIFY`, if neither `identity` nor `password` are set.
* `IDENTIFY password`, if only `password is set.
* `IDENTIFY identity`, if only `identity` is set. (Most nickserv services
  treat this identically to the previous case.)
* `IDENTIFY identity password`, if both `identity` and `password` are set.

This handler does not attempt to verify that it's actually talking to an
authorized nickserv service. Use with care.