Added support for invites.
If the bot receives an invite, and this handler is enabled, the bot will
attempt to join the channel it was invited to. These invitations are NOT
persistent; the bot makes a single attempt and does not retry, nor does it
tell anyone when it fails to join a channel, nor does it remember where it's
been invited to on restart. (If you want that, edit the conf file and update
the channel list there.)