Created an example upstart(8) configuration.
The example config presupposes that alicorn is installed in a virtualenv in
`/opt/alicorn`, that there is a user and group named `alicorn`, with
permissions in `/srv/alicorn`, and that there is a configuration file in
`/etc/alicorn.conf`. The bot is rigged to start once networking is up, and to
stop before networking goes away.

I've included a safety net `respawn` stanza; if, for some reason, the bot
exits abruptly, it'll be restarted -- but if it does so more than five times
in the same minute, it'll stop being restarted, to avoid wasting immense
amounts of CPU time looping over bot startup.