License under MIT.
This includes a simple contributor license agreement, meant to provide simple
demonstrations of contributors' willingness to licence their own code. The CLA
is _not_ meant to be legally bulletproof, and has not been vetted by a lawyer;
it's a simple best-effort attempt to forestall the problem and document who may
need to be contacted if legal issues do arise.