Imported module from a prior sketch.
`MybatisBundle` provides a fairly conventional module, broadly inspired by
`dropwizard-hibernate`'s `HibernateBundle`. It provides a session factory once
started, which can in turn be used to obtain mappers. We permit multiple
configuration styles, which overlap:

* Configure nothing.
* Configure an explicit set of mapper interfaces.
* Configure all mappers found in a given set of packages + subpackages.
* Apply arbitrary code to the `Configuration` object.

This should cover the majority of use cases: it certainly covers mine.

There's no support _yet_ for things like automated query decoration for
traceability or annotation-driven transaction demarcation. Those are probably
good ideas, but are considerably more complex to build and I wanted to get
something out fast to iterate on.

As is Right and Just, registering the bundle also registers a healthcheck for
the underlying database.