Publish builds to the login-box repos on Bintray.
As per usual, this makes it possible to use this library in other software.
I've included snapshot support.

The `maven-publish` plugin is declared late intentionally. It must be after the
`release{}` block, as interactions between `maven-publish` and the versioning
plugin we're using cause build failures if `maven-publish` looks at artefacts
(and thus versions) too soon:

    Inferred version: 0.0.1-dev.23+29cfdf1

    FAILURE: Build failed with an exception.

    * Where:
    Build file '/Users/owen/Development/login-box.com/dropwizard-mybatis/build.gradle' line: 69

    * What went wrong:
    A problem occurred evaluating root project 'dropwizard-mybatis'.
    > Cannot configure the 'publishing' extension after it has been accessed.

    * Try:
    Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output.

    BUILD FAILED