Released under the MIT license.
The `CONTRIBUTORS` directory is a best-effort attempt to document that
contributors understand that their code will be released under the same license
as the project itself. It has _not_ been vetted by a lawyer and is _not_ meant
to withstand serious legal scrutiny; it's meant to provide only the most minimal
documentation of consent.