Bake version info into the resulting JAR.
This improves traceability: given a bug report, you can now trace the binaries
back to a specific source version.

On Heroku, this relies on the `SOURCE_VERSION` environment variable. (See
https://devcenter.heroku.com/changelog-items/630.) This variable can also be
set locally to force a specific version string, but this is Not Wise.

In the absence of a `SOURCE_VERSION` environment variable, the build falls back
to getting version info from Git. To avoid failures on Heroku, `git` isn't even
invoked unless needed.