Display a nicer, more explanatory error when encountering CSRF problems.
This is meant to guide the _user_ to sensible behaviour, and to alert them to
the malicious or buggy site if possible. (Obviously, this won't be visible if
the CSRF attempt comes from an image tag or a script source, but we can try.)

This is separate from the exception itself so as to allow the CSRF bundle to be
severed from this project and released as a standalone library.