Define lifts for each transactable type.
Lifts convert non-transactable chunks of computation into transactable actions,
allowing them to be run in the context of a transaction. This also allows them
to be sequenced with other actions in a transaction, and to participate in
determining the final value of a transaction.

Providing explicit `lift` operations should help cut down on the need for dumb
code like

    Action<Context> action = context -> this.someVoidMethod();

in favour of

    Action<Context> action = lift(this::someVoidMethod);