MIT license the code, and create an example contributor agreement.
The contributor agreement is trivial, and merely restates the MIT license. The
point is not to be legally binding (I haven't involved a lawyer); it's to avoid
ambiguity as to the copyright and license terms of contributions from others
and to act as an indication of good faith as to the provenance of this code.