Specify buildpacks using `app.json` features, not env vars.
Heroku made it possible to do this
[recently](https://devcenter.heroku.com/changelog-items/670). It's not exposed
to deploying users the way `BUILDPACK_URL` is, which means the user can't screw
up their app by editing it to the wrong value (but can still set it to a
different value using the `heroku buildpacks:set` CLI command, if desired). It
also supports multiple buildpacks, but that's irrelevant here -- we're avoiding
that by dragging in Node via Gradle.