Pass URL parameters from `DATABASE_URL` through to JDBC.
This was motivated by ticket #32, wherein someone asked for support for
accessing Heroku databases from outside of Heroku. That's a reasonable need,
but because of the way Heroku manages ephemeral-ish hosts like database
servers, there's no good default way to use SSL: it always needs to be told to
disregard cert validation, which I refuse to ship as the default configuration,
and writing a custom `SSLSocketFactory` which recognizes Heroku's specific
setup is an unending maintenance effort.

This implementation makes it the user's problem, and documents it. It's got
some ugly wrinkles around URL parameters with escapable characters in them, but
I think those will be rare-to-nonexistant in practice.