Hoist directory ownership out to its own bundle.
This is prep work for the login bundle, which will need to share the directory
system.

I've re-applied Coda Hale's idiom for dependency injection using abstract
methods here, which makes `Bootstrap` feel a tiny bit inconsistent, but it's
probably the smoothest way to late-bind a bundle-provided dependency for a
component like `Bootstrap`. We'll see.

The Transactor system is a bloody hazard to refactoring.