Defined "pivot" operators.
Pivots transpose the arguments of a transactable with the context, producing a
new transactable whose context was the original's argument (and vice versa).
The primary motivation is to enable the use of method handles as transactables
when the handle may have been obtained via the class, rather than via a
specific instance.

For a complete description, see the changes to `README.md`.