Added an @Incomplete annotation that allows acceptance tests to be written early.
This supports the workflow where an acceptance test is written out before the
feature is fully realized; marking the test with `@Incomplete` means that it
will no longer cause build failures if tests are run during builds, but will
still exercise the feature (unlike JUnit's native `@Ignore`, which skips the
test outright).

If the test passes, the test will be _reported_ as a failure, to encourage
developers to review features they implement and to keep tests up to date.