Return a static HTML view on all routes.
It doesn't exist _yet_, but this will support HTML5 client-side routing once we
have a client app.

This is also the first inkling of the bundle system I have in mind. The gist is
that every major functional block gets its own bundle, so that dependencies can
be kept simple (or at least kept visible).