Split out a vendor chunk.
With chunk hashes, we can now aggressively cache chunks when they don't change.
I've set up Webpack to exploit this. The Webpack runtime lives in the current
`boot.[chunkhash].js` chunk, with libs in `vendor.[chunkhash].js` and app code
in `app.[chunkhash].js`. The boot chunk is tiny, and contains some metadata for
the other chunks.

* When app code changes, the `app` and `boot` chunks change.
* When vendor code changes (upgrades, new libs, etc), the `vendor` and `boot`
  chunks change.

The vendor chunk is the vast majority of the code in Distant Shore.