Use Auth0-Lock from Auth0's CDN.
The `auth0-lock` NPM package has (transitive) git dependencies, which fail to
install on Aerobatic. Rely on Auth0's own builds, instead, and make a
reasonable effort to check that we're receiving the code we expect. (It's a
token effort; subresource integrity checks don't work on many popular browsers,
though they do work on Chrome and Firefox.)

On the upside, this chops around a megabyte of code out of the vendor bundle.
The CDN asset is about 770kb, but can be downloaded in parallel with the other
bundles.