Generate config from `npm local`.
Moving off of Aerobatic means we now have to provide our own config system.
I've chosen to keep the style, if not necessarily the exact API; as a result,
we now generate a config script (which assigns a JS global) inside of Express.

`npm local` is now the canonical server. Ask me how I feel about this.