Break up boot steps.
I'd like to be able to boot from a locally-stored token, not just from a window
hash. To do this, I need a clearer place to insert alternate boot paths.

I've also cleaned up the Auth0 profile fetch to use promises, so that I can
more easily do it in parallel with the API whoami check.