Refresh expiring tokens.
I've picked "half the remaining duration" more or less out of a hat (loosely
inspired by DNS cache behaviour). In most cases, this attempts the refresh
_well_ before the token may expire, which helps ensure the refresh doesn't fail
due to happening too late.

The actual refresh process is fairly dumb: we try once; if it fails, we assume
the current token is fucked and throw it out (which is probably pessimistic),
regardless of cause.

There should only be at most one refresh timer going at a time, but the timer
lifecycle is not mediated by an action or reducer.