Require a PostgreSQL database.
We'll be building some persistent features, such as stored squads. To that end,
this commit builds in basic support for PostgreSQL, via a dedicated bundle.

The bundle does its own health-check management, because we're going to be
working directly with JDBC in places (mainly, in Axon) rather than solely
through a higher-level framework like JDBI or Mybatis.