Apply database migrations.
If you have a database, sooner or later you're going to need to create a schema
in it. Dropwizard Migrations (liquibase, really) is a proven, effective way to
do that.

This change includes Heroku configuration to apply database migrations on
release. While running migrations during deploy (when they're hard to
interrupt) isn't optimal, having to manage them manually is worse.