Use Axon for domain modeling.
Axon provides a rigid but complete implementation of CQRS and event sourcing
concepts, appropriate for constructing a game on top of. Provide a bundle that
sets up Axon.

I've configured it to use the marginally-terser JSON serialization, rather than
the default XStream XML serialization, and to store data as bare JSON strings
(rather than `bytea` encodings) in the database, for ease of reading.