Load polyfills from the vendor bundle.
After reading the code more closely, I realized the polyfills only have to be
executed to be loaded. Since the vendor bundle executes every top-level
dependency as a side effect of the way its contents are specified, adding the
polyfills to `dependencies` is enough to have them applied to the page.

The app bundle now contains only the app code.