Moved boot logic out of individual components.
By creating a route factory, rather than creating a simple route object, it's
now possible to inject dependencies into the router. This means that, for
example, route enter and leave hooks can interact with the Redux store and with
the API.

This simplifies individual components by quite a bit, as they no longer rely on
having access to the API _on mount_ and can instead be mounted safely in
contexts where the API is unavailable.