Fall back to Received: headers when Date: is missing or garbage.
This adds some robustness to message date detection, without altering explicit
date specification in the top-level API. The motivating case was an email in
agora-business, whose headers include

    Date: some time near the end of July

The Received headers were sufficient to adequately work out _when_ in July, or
close enough to.

This has the potential to cause some ordering anomalies if the timezone of a
mailer and the timezone of the last hop are seriously misaligned. I believe
submitting a revision with an explicit date is the right way to fix that, so no
code is needed.