Fix some O(n) query behaviours.
In staging, I noticed that the document list was multiple seconds slow. This surprised me, as it's only a few thousand rows. Turns out it was emitting one select for the list of documents, and then one select per document for the list of revisions, and then one select per document for the list of annotations. Erm.

Eagerly loading these causes the initial select to return a lot more data, but cuts down the round-tripping to the DB.