Simplify and unify the output formats of the list_documents and retrieve_document endpoints.
The `list_documents` output was actually broken (the `url`, `download_url`, and `annotation_url` fields were being omitted from the output for reasons unknown), but removing the distinction between formats is more consistent for users.

This does slow down `list_documents`, as it now fetches every annotation and revision in the DB. See #4 for details on this.