Save states to the DB in the background, dropping any states where the previous save took too long.
This can leave gaps in the list of recorded saves. That's fine. The code will cope by reconstructing missing states from events.

I also rewrote the client catch-up logic to deal with cases where a client has remained disconnected for so long that their event ID is too stale to recover from. We no longer send a state re-sync unless the client has fallen out of range, but if the client has fallen out of range, we re-sync to the oldest available state and carry on as if nothing happened.