Added a security warning to the README.
I don't intend to even _attempt_ to audit all of this code for security issues, nor do I have the time or bandwidth to stay on top of package updates. This archive is clearly labelled as a dusty old repository full of sharp edges, and it's the best I can do while keeping the code online.

As I do still sometimes refer to this code, I'm not going to delete it or lock it down at this time.