A websockets-based chat system to use as a learning example.
The chat system is very simple: any user can send a message to all connected users, claiming to be an username and with any content they want (so long as it's text). It's based on WebSockets, since that's what's easily available.

I wrote this for a friend of mine as a favour, so that she could use it to run her class through their paces writing chat bots.