Retire the MIT license from my code.
I no longer believe it's in my best interests to allow anyone, anywhere
to build on top of my work without so much as talking to me. I still
believe strongly that the best code is built on the work of others, and
that sharing code is an important building block for equitable and
humane computing, but the MIT license doesn't adequately differentiate
between those goals and, say, taking my code to build a product.

The `unacceptable` project remains MIT-licensed, as I haven't discussed
relicensing it with the other contributors. Code imported from other
projects also continues to hold those projects' licenses.